# Improving sync parachain performance for workers