--- a/tee-worker/cli/lit_test_failed_parentchain_extrinsic.sh
+++ b/tee-worker/cli/lit_test_failed_parentchain_extrinsic.sh
@@ -46,7 +46,7 @@ echo "New Account created: ${FIRST_NEW_ACCOUNT}"
 echo "Linking identity to Bob"
 OUTPUT=$(${CLIENT} link-identity //Bob did:litentry:substrate:${FIRST_NEW_ACCOUNT} litentry) || { echo "Link identity command failed"; exit 1; }
 echo "Finished Linking identity to Bob"
-sleep 30
+sleep 60
 
 echo "Capturing IDGraph Hash of Bob" 
 INITIAL_ID_GRAPH_HASH=$(${CLIENT} id-graph-hash did:litentry:substrate:0x8eaf04151687736326c9fea17e25fc5287613693c912909cb226aa4794f26a48) || { echo "Failed to get ID Graph hash"; exit 1; }
@@ -58,7 +58,7 @@ echo "New Account created: ${SECOND_NEW_ACCOUNT}"
 echo "Linking new identity to Bob with Eve as delegate signer"
 OUTPUT=$(${CLIENT} link-identity //Bob "did:litentry:substrate:${SECOND_NEW_ACCOUNT}" litentry -d //Eve) || { echo "Link identity command failed"; exit 1; }
 echo "Finished Linking identity to Bob"
-sleep 30
+sleep 60
 
 echo "Capturing IDGraph Hash of Bob" 
 FINAL_ID_GRAPH_HASH=$(${CLIENT} id-graph-hash did:litentry:substrate:0x8eaf04151687736326c9fea17e25fc5287613693c912909cb226aa4794f26a48) || { echo "Failed to get ID Graph hash"; exit 1; }