This PR improves the performance for sync with parachains a bit. I've tested it with litentry rococo. 

Here are the results:

Current:
```
[2024-07-11T15:03:34.444Z INFO  litentry_worker::parentchain_handler] [Litentry] Syncing blocks from 0 to 24043
[2024-07-11T15:03:34.444Z DEBUG litentry_worker::parentchain_handler] >>>>>>>>>>>>>>>>>>>> Starting batch sync <<<<<<<<<<<<<<<<<<<<
[2024-07-11T15:03:34.909Z INFO  litentry_worker::parentchain_handler] [Litentry] syncing parentchain to 24043. already synced until block 0. immediate import=true. est. remaining: 3h 6m 3s 526ms 747us 150ns
[2024-07-11T15:03:34.909Z DEBUG litentry_worker::parentchain_handler] [Litentry] Fetched 1001 blocks in 464ms 264us 889ns
[2024-07-11T15:03:38.867Z DEBUG litentry_worker::parentchain_handler] >>>>> [Litentry] Synced parentchain batch in 3s 958ms 550us 433ns
[2024-07-11T15:03:38.867Z INFO  litentry_worker::parentchain_handler] [Litentry] Synced 1001 out of 24043 finalized parentchain blocks
[2024-07-11T15:03:39.316Z DEBUG litentry_worker::parentchain_handler] [Litentry] Fetched 1001 blocks in 447ms 820us 558ns
[2024-07-11T15:03:46.684Z DEBUG litentry_worker::parentchain_handler] >>>>> [Litentry] Synced parentchain batch in 7s 368ms 798us 150ns
[2024-07-11T15:03:46.684Z INFO  litentry_worker::parentchain_handler] [Litentry] Synced 2002 out of 24043 finalized parentchain blocks
[2024-07-11T15:03:47.124Z DEBUG litentry_worker::parentchain_handler] [Litentry] Fetched 1001 blocks in 439ms 177us 262ns
[2024-07-11T15:03:54.511Z DEBUG litentry_worker::parentchain_handler] >>>>> [Litentry] Synced parentchain batch in 7s 386ms 474us 448ns
[2024-07-11T15:03:54.511Z INFO  litentry_worker::parentchain_handler] [Litentry] Synced 3003 out of 24043 finalized parentchain blocks
[2024-07-11T15:03:54.959Z DEBUG litentry_worker::parentchain_handler] [Litentry] Fetched 1001 blocks in 447ms 534us 184ns
[2024-07-11T15:04:02.491Z DEBUG litentry_worker::parentchain_handler] >>>>> [Litentry] Synced parentchain batch in 7s 532ms 744us 764ns
[2024-07-11T15:04:02.492Z DEBUG litentry_worker::parentchain_handler] [Litentry] Synced 4004 out of 24043 finalized parentchain blocks

```

New: 

```
[2024-07-11T14:41:33.548Z INFO  litentry_worker::parentchain_handler] [Litentry] Syncing blocks from 0 to 23933
[2024-07-11T14:41:33.548Z DEBUG litentry_worker::parentchain_handler] >>>>>>>>>>>>>>>>>>>> Starting batch sync <<<<<<<<<<<<<<<<<<<<
[2024-07-11T14:41:33.635Z INFO  litentry_worker::parentchain_handler] [Litentry] syncing parentchain to 23933. already synced until block 0. immediate import=true. est. remaining: 34m 23s 468ms 389us 619ns
[2024-07-11T14:41:33.635Z DEBUG litentry_worker::parentchain_handler] [Litentry] Fetched 1000 blocks in 86ms 182us 830ns
[2024-07-11T14:41:33.635Z DEBUG litentry_worker::parentchain_handler] [Litentry] Found 1000 block(s) to sync in this chunk. immediate import=true
[2024-07-11T14:41:37.682Z DEBUG litentry_worker::parentchain_handler] >>>> [Litentry] Synced parentchain batch in 4s 47ms 236us 657ns
[2024-07-11T14:41:37.682Z INFO  litentry_worker::parentchain_handler] [Litentry] Synced 1000 out of 23933 finalized parentchain blocks
[2024-07-11T14:41:37.765Z DEBUG litentry_worker::parentchain_handler] [Litentry] Fetched 1000 blocks in 82ms 435us 816ns
[2024-07-11T14:41:37.765Z DEBUG litentry_worker::parentchain_handler] [Litentry] Found 1000 block(s) to sync in this chunk. immediate import=true
[2024-07-11T14:41:45.298Z DEBUG litentry_worker::parentchain_handler] >>>> [Litentry] Synced parentchain batch in 7s 533ms 92us 133ns
[2024-07-11T14:41:45.298Z INFO  litentry_worker::parentchain_handler] [Litentry] Synced 2000 out of 23933 finalized parentchain blocks
[2024-07-11T14:41:45.384Z DEBUG litentry_worker::parentchain_handler] [Litentry] Fetched 1000 blocks in 85ms 803us 899ns
[2024-07-11T14:41:45.384Z DEBUG litentry_worker::parentchain_handler] [Litentry] Found 1000 block(s) to sync in this chunk. immediate import=true
[2024-07-11T14:41:52.704Z DEBUG litentry_worker::parentchain_handler] >>>> [Litentry] Synced parentchain batch in 7s 319ms 210us 969ns
[2024-07-11T14:41:52.704Z INFO  litentry_worker::parentchain_handler] [Litentry] Synced 3000 out of 23933 finalized parentchain blocks
[2024-07-11T14:41:52.789Z DEBUG litentry_worker::parentchain_handler] [Litentry] Fetched 1000 blocks in 84ms 480us 849ns
[2024-07-11T14:41:52.789Z DEBUG litentry_worker::parentchain_handler] [Litentry] Found 1000 block(s) to sync in this chunk. immediate import=true
[2024-07-11T14:42:00.305Z DEBUG litentry_worker::parentchain_handler] >>>> [Litentry] Synced parentchain batch in 7s 516ms 377us 787ns
[2024-07-11T14:42:00.305Z INFO  litentry_worker::parentchain_handler] [Litentry] Synced 4000 out of 23933 finalized parentchain blocks
[2024-07-11T14:42:00.391Z DEBUG litentry_worker::parentchain_handler] [Litentry] Fetched 1000 blocks in 85ms 436us 539ns
[2024-07-11T14:42:00.391Z DEBUG litentry_worker::parentchain_handler] [Litentry] Found 1000 block(s) to sync in this chunk. immediate import=true
[2024-07-11T14:42:07.858Z DEBUG litentry_worker::parentchain_handler] >>>> [Litentry] Synced parentchain batch in 7s 467ms 196us 715ns
[2024-07-11T14:42:07.858Z INFO  litentry_worker::parentchain_handler] [Litentry] Synced 5000 out of 23933 finalized parentchain blocks
```
Notice how the time to fetch the blocks is significantly reduced by calling the api in parallel. There may be some more room for improvement but we need to measure first to be able to find the bottlenecks, which is not easy task on development when we don't have the same data as in prod.

In the future (once we migrate to the new sgx sdk), we could use the async version of the substrate-api-client. More context in the issue.