@Kailai-Wang 2024-07-15
> Running `lit_test_failed_parentchain_extrinsic.sh` works for me locally, I'm not sure what the problem is with CI
> 
> ```
> Using client binary ./../bin/litentry-cli
> Using node uri ws://127.0.0.1:10044
> Using trusted-worker uri ws://127.0.0.1:2000
> 
> CLIENT is: ./../bin/litentry-cli -p 10044 -P 2000 -u ws://127.0.0.1 -U ws://127.0.0.1
> New Account created: 0xd266a81c0d51c9b0ece2f85daa3d659e4bde596d2fd42d555109730e70255f0f
> Linking identity to Bob
> Finished Linking identity to Bob
> Capturing IDGraph Hash of Bob
> Initial ID Graph Hash of Bob: 0x8cffcda6d69f27ba44bf82adf4b9a244b0c1560c2c057280f84c15ea381c2818
> New Account created: 0x8c4f7ca222e5e30323979918435d7e77899130ee0923a2d30f76510580b13155
> Linking new identity to Bob with Eve as delegate signer
> Finished Linking identity to Bob
> Capturing IDGraph Hash of Bob
> Final ID Graph Hash of Bob: 0x8cffcda6d69f27ba44bf82adf4b9a244b0c1560c2c057280f84c15ea381c2818
> Failed Parentchain extrinsic not executed
> ```

Hmm if it only fails in this PR while working elsewhere, we need to look into it