@silva-fj 2024-07-15
> Hmm if it only fails in this PR while working elsewhere, we need to look into it

I can't reproduce it on `azure-dev-2`, I'll add some logs to see if I can find something on CI