@felixfaisal 2024-07-16
> Can you elaborate on this? Ideally we should be using the async api to make these requests but unfortunately we can't use it yet. Nevertheless, calling these blocking functions in parallel does help to improve the sync a bit

Yeah I would prefer async over multi-threading since these are mostly I/O bound tasks, but I guess multi-threading is the only option we currently have. Something else related to infrastructure is if the node can support so many requests in prod without terminating the WS connection but this is out of scope for this PR. 

> please note that this is used during initialisation, so I think it should be fine if it takes some resources

I missed out on this, then in that case, i guess it's fine to allocate as much as possible 

 > I think we should probably wait until the initialisation is done (maybe calling the /is_initialized endpoint) before calling the cli commands in the test.
 
 Sounds like a good idea