@silva-fj 2024-07-16
I've spent quite some time debugging on CI and trying to make the script to work by calling the `/is_initialized` endpoint and running the tests after the initialization but I could not make it work (on CI), maybe I was not using the right port. So, for now I've increased the sleep time before requesting the id graphs to make the test pass