@Kailai-Wang 2024-07-18
I'm ok-ish with it, although I'm not sure how many CPUs will be available within an enclave.

Can you test it a bit with sdk-v2.0 branch before merging it?
I ask this because we have a lower number of TCSNum there