@silva-fj 2024-07-22
> I'm ok-ish with it, although I'm not sure how many CPUs will be available within an enclave.
> 
> Can you test it a bit with sdk-v2.0 branch before merging it? I ask this because we have a lower number of TCSNum there

There is no change in how the enclave api is being called, it's still 1 call at a time on each loop. What is different now is how we collect the data to pass to the enclave api on each loop