@silva-fj 2024-07-22
> Still a bit skeptical about how this would work in prod or with a live chain

If our production nodes can't handle this load, we have bigger problems that this 😆